+++ ortto_flutter_sdk_platform_interface/lib/ortto_flutter_sdk_platform_interface.dart
@@ -81,4 +81,3 @@ abstract class OrttoFlutterSdkPlatformInterface extends PlatformInterface {
 
   Future<IdentityResult> clearIdentity();
 }
-