+++ ortto_flutter_sdk_platform_interface/lib/src/models/ortto_config.dart
@@ -2,23 +2,18 @@ class OrttoConfig {
   String appKey;
   String endpoint;
   bool shouldSkipNonExistingContacts = false;
-  bool allowAnonUsers = false;
 
   OrttoConfig(
     this.appKey,
     this.endpoint, {
-      this.shouldSkipNonExistingContacts = false,
-      this.allowAnonUsers = false,
-    }
-  );
+    this.shouldSkipNonExistingContacts = false,
+  });
 
   Map<String, dynamic> toMap() {
     return {
       'appKey': appKey,
       'endpoint': endpoint,
       'shouldSkipNonExistingContacts': shouldSkipNonExistingContacts,
-      'allowAnonUsers': allowAnonUsers,
     };
   }
 }
-