+++ ortto_flutter_sdk_ios/ios/ortto_flutter_sdk_ios/Sources/ortto_flutter_sdk_ios/OrttoFlutterSdkPlugin.swift
@@ -4,7 +4,6 @@ import OrttoSDKCore
 import OrttoPushMessagingFCM
 import OrttoInAppNotifications
 import OrttoPushMessaging
-import FirebaseMessaging
 
 public class OrttoFlutterSdkPlugin: NSObject, FlutterPlugin, UNUserNotificationCenterDelegate {
     public static func register(with registrar: FlutterPluginRegistrar) {